refactor: external 패키지 생성
RandomNumbersGenerator를 외부 의존성을 관리하는
external 패키지로 이동한다.
