+++ src/main/java/lotto/Application.java
@@ -1,13 +1,14 @@
 package lotto;
 
 import lotto.domain.Lottos;
-import lotto.domain.RandomNumbersGenerator;
+import lotto.external.RandomNumbersGenerator;
 
 public class Application {
 
     public static void main(String[] args) {
 
         Lottos lottos = new Lottos(3, new RandomNumbersGenerator());
+
 
     }
 }
